move tenant to internal package
