+++ backend/app/instance_provider.go
@@ -6,7 +6,7 @@ import (
 
 	"github.com/grafana/grafana-plugin-sdk-go/backend"
 	"github.com/grafana/grafana-plugin-sdk-go/backend/instancemgmt"
-	"github.com/grafana/grafana-plugin-sdk-go/backend/tenant"
+	"github.com/grafana/grafana-plugin-sdk-go/internal/tenant"
 )
 
 // InstanceFactoryFunc factory method for creating app instances.
